+++ lib/que/scheduler/migrations/1/down.sql
@@ -1,0 +1,1 @@
+DELETE FROM que_jobs WHERE job_class = 'Que::Scheduler::SchedulerJob';