+++ lib/que/scheduler/migrations/1/up.sql
@@ -1,0 +1,1 @@
+-- Do nothing. This is handled by an "enqueue" call.