+++ lib/que/scheduler/migrations/2/down.sql
@@ -1,0 +1,1 @@
+DROP TABLE que_scheduler_audit;