+++ spec/que/scheduler/db_spec.rb
@@ -1,0 +1,20 @@
+require 'spec_helper'
+
+RSpec.describe Que::Scheduler::Db do
+  describe '.count_schedulers' do
+    it 'returns the right result' do
+      expect(described_class.count_schedulers).to eq(0)
+      ::Que::Scheduler::SchedulerJob.enqueue
+      expect(described_class.count_schedulers).to eq(1)
+      ::Que::Scheduler::SchedulerJob.enqueue
+      expect(described_class.count_schedulers).to eq(2)
+    end
+  end
+
+  describe '.now' do
+    it 'returns the time' do
+      expect(Que).to receive(:execute).with(described_class::NOW_SQL).and_return([{ foo: :bar }])
+      expect(described_class.now).to eq(:bar)
+    end
+  end
+end